% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSRplot.R
\name{blimprior}
\alias{blimprior}
\title{blimprior plot
Plots the bounds of the hockey-stick break-point}
\usage{
blimprior(
  lplim = 0.01,
  uplim = 0.3,
  s.logitsd = 50,
  par = "plim",
  bias.correct = TRUE
)
}
\arguments{
\item{lplim}{steepness, default 0.6 for a approx. uniform prior with s.logistsd = 20}

\item{uplim}{s steepness, default 20 for a approx. uniform prior with s = 0.6}

\item{par}{parameter on x-axis default "plim", else "sstar"}

\item{s.sdlogit}{default 20}
}
\value{
ggplot
}
\description{
blimprior plot
Plots the bounds of the hockey-stick break-point
}
\examples{
blimprior() # approx. uniform with some curving on bounds
blimprior(lplim=0.001,uplim=0.3,s.logitsd=20) 
# Non-bias corrected
blimprior(lplim=0.001,uplim=0.3,s.logitsd=20,bias.correct=FALSE)
}
