% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSRplot.R
\name{blimprior1}
\alias{blimprior1}
\title{blimprior1 plot
Plots the bounds of the hockey-stick break-point}
\usage{
blimprior1(lplim = 0.05, uplim = 0.2, s.logitsd = 20, par = "plim")
}
\arguments{
\item{lplim}{steepness, default 0.6 for a approx. uniform prior with s.logistsd = 20}

\item{uplim}{s steepness, default 20 for a approx. uniform prior with s = 0.6}

\item{par}{parameter on x-axis default "plim", else "sstar"}

\item{s.sdlogit}{default 20}
}
\value{
ggplot
}
\description{
blimprior1 plot
Plots the bounds of the hockey-stick break-point
}
\examples{
blimprior1(lplim=0.001,uplim=0.2) # approx. uniform with some curving on bounds
}
