% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrapSR}
\alias{bootstrapSR}
\title{Bootstrap fits of mutliple stock-recruits relationships}
\usage{
bootstrapSR(
  x,
  iters = 100,
  method = c("best", "logLik", "relative"),
  models = c("bevholt", "ricker", "segreg"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class 'FLStock'.}

\item{models}{Name(s) of model(s) to fit, 'character'. See Details.}

\item{verbose}{Should progress be reported, 'logical'.}

\item{iter}{Number of bootstrap iterations, 'numeric'.}
}
\value{
A list with elements An object or class 'FLPar' containing the estimated paramaters.
}
\description{
Definition ...
}
\details{
The returned 'FLPar' object contains
}
\examples{
data(ple4)
mods <- bootstrapSR(ple4, iter=50, model=c("bevholt", "segreg"))
}
\seealso{
\link{FLSR}, link{srrTMB}
}
\author{
Iago Mosqueira (WMR)
}
\keyword{models}
