% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSRplot.R
\name{dprior}
\alias{dprior}
\title{dprior plot
Plots the logit prior distribution for depensation}
\usage{
dprior(d = 1, d.logitsd = 100, ll = 0.5, ul = 3)
}
\arguments{
\item{d}{depensation, default 1 for a approx. uniform prior with s.logistsd = 20}

\item{d.logitsd}{dependation sd, default 20 for a approx. uniform prior with s = 0.6}

\item{ll}{lower bound of d = 0.25}

\item{ul}{lower bound of d = 4}
}
\value{
ggplot
}
\description{
dprior plot
Plots the logit prior distribution for depensation
}
\examples{
dprior() # approx. uniform with some curving on bounds
dprior(d=1,d.logitsd=2) 
}
