% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSRTMB-package.R
\docType{package}
\name{FLSRTMB-package}
\alias{FLSRTMB}
\alias{FLSRTMB-package}
\title{FLSRTMB: FLSR in TMB}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Estimates FLR spawner recruitment relationships in TMB
}
\author{
\strong{Maintainer}: Henning Winker \email{henning.winker@ec.europa.eu}

Authors:
\itemize{
  \item Iago Mosqueira \email{iago.mosqueira@wur.nl} (\href{https://orcid.org/0000-0002-3252-0591}{ORCID})
}

Other contributors:
\itemize{
  \item European Union [funder, copyright holder]
  \item Christoffer Moesgaard Albertsen \email{cmoe@aqua.dtu.dk} (\href{https://orcid.org/0000-0003-0088-4363}{ORCID}) [contributor]
}

}
\keyword{internal}
