% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_logits}
\alias{from_logits}
\title{from_logits()}
\usage{
from_logits(logit_s, ll = 0.2, ul = 1)
}
\arguments{
\item{logit_s}{logit(steepness)}

\item{ll}{defines lower prior bound of s}

\item{ul}{defines upper prior bound of s}
}
\value{
steepness s
}
\description{
convert steepness from logit
}
