% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hsblim}
\alias{hsblim}
\title{hsblim()}
\usage{
hsblim(object)
}
\arguments{
\item{object}{class FLSR fitted with model segreg}
}
\value{
FLPar
}
\description{
helper to quickly extract blim, r0 and srp0 = blim/b0 from hockeystick
}
