% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jitter.R
\name{plotjitter}
\alias{plotjitter}
\title{plots Jitter results}
\usage{
plotjitter(jitter)
}
\arguments{
\item{jitter}{output from srrjitter()}
}
\value{
ggplot
}
\description{
plots Jitter results
}
\examples{
data(ple4)
hs = srrTMB(as.FLSR(ple4,model=segreg),spr0=spr0y(ple4),lplim=0.07,uplim=0.2)
plotsrs(hs)
jitter = srrjitter(hs)
plotsrs(jitter$groups)
plotjitter(jitter)
bh = srrTMB(as.FLSR(ple4,model=bevholtSV),spr0=spr0y(ple4))
jitbh = srrjitter(bh)
plotjitter(jitbh)
plotsrs(jitbh$groups)
}
