% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSRplot.R
\name{plotsrs}
\alias{plotsrs}
\title{plotsrs {{{ based on plot(FLSRs)
Plots FLSRs, i.e. multiple S-R relationships}
\usage{
plotsrs(object, path = TRUE)
}
\arguments{
\item{object}{of class FLSRS}

\item{path}{connect points sequentially}
}
\value{
ggpplot
}
\description{
plotsrs {{{ based on plot(FLSRs)
Plots FLSRs, i.e. multiple S-R relationships
}
