% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSRplot.R
\name{plotsrts}
\alias{plotsrts}
\title{Plots time series of observed and predicted recruitment}
\usage{
plotsrts(object)
}
\arguments{
\item{object}{Input FLSR of FLSRs object}
}
\value{
ggplot
}
\description{
Plots time series of observed and predicted recruitment
}
\examples{
data(ple4)
bh <- srrTMB(as.FLSR(ple4,model=bevholtSV),spr0=mean(spr0y(ple4)))
plotsrts(bh)
# Try more models
hs <- srrTMB(as.FLSR(ple4,model=segreg),spr0=(spr0y(ple4)),plim=0.02,pmax=0.3)
ri <- srrTMB(as.FLSR(ple4,model=rickerSV),spr0=mean(spr0y(ple4)))
bh.tv <- srrTMB(as.FLSR(ple4,model=bevholtSV),spr0=spr0y(ple4))
ri.tv <- srrTMB(as.FLSR(ple4,model=rickerSV),spr0=spr0y(ple4))
all= FLSRs(hs=hs,bh=bh,ri=ri,bh.tv=bh.tv,ri.tv=ri.tv)
plotsrts(all)
do.call(c,lapply(all,function(x)AIC(x))) # AIC
plot(all)+theme(legend.position="right")
}
