% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resilience}
\alias{resilience}
\title{resilience()}
\usage{
resilience(object, s = 0.7)
}
\arguments{
\item{object}{class FLStock}

\item{s}{steepness of the stock recruitment relationship}
}
\value{
FLQuants with FLQuant r, gt and alpha
}
\description{
Function to compute r, generation time (gt) and annual reproductive rate (alpha)
}
\author{
Henning Winker and Laurence Kell
}
