% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spr0y}
\alias{spr0y}
\title{spr0y()}
\usage{
spr0y(object, byage = FALSE)
}
\arguments{
\item{object}{class FLStock}

\item{byage}{if TRUE it return spr0_at_age}
}
\value{
FLQuant with annual spr0y
}
\description{
Function to compute annual spr0
}
\author{
Laurence Kell
}
