% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sprFy}
\alias{sprFy}
\title{sprFy()}
\usage{
sprFy(object, byage = FALSE)
}
\arguments{
\item{object}{class FLStock}

\item{byage}{if TRUE it return sprF_at_age}
}
\value{
FLQuant with annual sprFy
}
\description{
Function to compute annual sprF as function of F_a
}
\author{
Henning Winker and Laurence Kell
}
