% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSRplot.R
\name{sprior}
\alias{sprior}
\title{sprior plot
Plots the logit prior distribution for steepness}
\usage{
sprior(s = 0.6, s.logitsd = 20, ll = 0.2, ul = 1)
}
\arguments{
\item{s}{steepness, default 0.6 for a approx. uniform prior with s.logistsd = 20}

\item{s.logitsd}{s steepness, default 20 for a approx. uniform prior with s = 0.6}

\item{ll}{lower bound of s = 0.2}

\item{ul}{lower bound of s = 1}
}
\value{
ggplot
}
\description{
sprior plot
Plots the logit prior distribution for steepness
}
\examples{
sprior() # approx. uniform with some curving on bounds
sprior(s=0.8,s.logitsd=0.5) 
}
