% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jitter.R
\name{srrjitter}
\alias{srrjitter}
\title{Jitter of S-R fits}
\usage{
srrjitter(fit, steps = 100)
}
\arguments{
\item{fit}{fit of srrTBM()}

\item{steps}{number of jitter steps}
}
\value{
list
}
\description{
Jitter of S-R fits
}
\examples{
data(ple4)
hs = srrTMB(as.FLSR(ple4,model=segreg),spr0=spr0y(ple4),lplim=0.07,uplim=0.5)
plotsrs(hs)
jitter = srrjitter(hs)
plotsrs(jitter$groups)
plotsrs(jitter$best) # Best
}
