% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSRtest.R
\name{srrTMB1}
\alias{srrTMB1}
\title{Fits Stock Recruitment Relationships (SRR) in TMB}
\usage{
srrTMB1(object, ...)
}
\arguments{
\item{object}{Input FLSR = as.FLSR(stock,model) object with current model options
\itemize{
\item bevholtSV
\item bevholtDa
\item rickerSV
\item segreg
\item geomean
}}

\item{s}{steepness parameter of SRR (fixed or prior mean)}

\item{spr0}{unfished spawning biomass per recruit from FLCore::spr0(FLStock)}

\item{s.est}{option to estimate steepness}

\item{s.logitsd}{prior sd for logit(s), default is 1.4 (flat) if s.est = TRUE}

\item{r0.pr}{option to condition models on r0 priors (NULL = geomean)}

\item{lplim}{lower bound of spawning ratio potential SRP, default 0.0001}

\item{uplim}{upper bound of plausible spawning ratio potential SRP , default 0.3}

\item{Blim}{fixing Blim, only works with segreg}

\item{nyears}{yearMeans from the tail used to compute a,b from the reference spr0 (default all years)}

\item{report.sR0}{option to report s and R0 instead of a,b}

\item{inits}{option to specify initial values of log(r0), log(SigR) and logit(s)}

\item{lower}{option to specify lower bounds of log(r0), log(SigR) and logit(s)}

\item{upper}{option to specify upper bounds of log(r0), log(SigR) and logit(s)}

\item{SDreport}{option to converge hessian and get vcov}

\item{rm.yrs}{remove recruitment years from model fit}

\item{verbose}{if TRUE, it shows tracing}
}
\value{
A list containing elements 'FLSR', of class \emph{FLSR}
}
\description{
Fits Stock Recruitment Relationships (SRR) in TMB
}
\examples{
data(ple4)
gm <- srrTMB1(as.FLSR(ple4,model=geomean),spr0=mean(spr0y(ple4)))
bh <- srrTMB1(as.FLSR(ple4,model=bevholtSV),spr0=spr0y(ple4))
ri <- srrTMB1(as.FLSR(ple4,model=rickerSV),spr0=spr0y(ple4))
hs <- srrTMB1(as.FLSR(ple4,model=segreg),spr0=spr0y(ple4),lplim=0.001,uplim=0.2)
srs = FLSRs(gm=gm,bh=bh,ri=ri,hs=hs) # combine
plotsrs(srs) 
plotsrts(srs)  # relative
plotsrs(srs[2:4],b0=TRUE) # through to B0
plotsrs(srs[2:4],b0=TRUE,rel=TRUE)  # relative
gm@SV # estimates
do.call(rbind,lapply(srs,AIC))
}
