% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_logitd}
\alias{to_logitd}
\title{to_logitd()}
\usage{
to_logitd(d, ll = 0.5, ul = 3)
}
\arguments{
\item{d}{steepness s}

\item{ll}{defines lower prior bound of s}

\item{ul}{defines upper prior bound of s}
}
\value{
logit transformed steepness s
}
\description{
convert steepness to logit
}
