% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_logits}
\alias{to_logits}
\title{to_logits()}
\usage{
to_logits(s, ll = 0.2, ul = 1)
}
\arguments{
\item{s}{steepness s}

\item{lim}{defines lower prior bound of s}

\item{max}{defines upper prior bound of s}
}
\value{
logit transformed steepness s
}
\description{
convert steepness to logit
}
