% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSRplot.R
\name{blimprior}
\alias{blimprior}
\title{blimprior plot
Plots the bounds of the hockey-stick break-point}
\usage{
blimprior(lplim = 0.05, uplim = 0.2, s.logitsd = 10, par = "plim")
}
\arguments{
\item{lplim}{steepness, default 0.6 for a approx. uniform prior with s.logistsd = 20}

\item{uplim}{s steepness, default 20 for a approx. uniform prior with s = 0.6}

\item{par}{parameter on x-axis default "plim", else "sstar"}

\item{s.sdlogit}{default 20}
}
\value{
ggplot
}
\description{
blimprior plot
Plots the bounds of the hockey-stick break-point
}
\examples{
sprior() # approx. uniform with some curving on bounds
sprior(s=0.8,s.logitsd=0.5) 
}
