% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_logits}
\alias{from_logits}
\title{from_logits()}
\usage{
from_logits(logit_s, lim = 0.2)
}
\arguments{
\item{logit_s}{logit(steepness)}

\item{lim}{defines lower prior bound of s}
}
\value{
steepness s
}
\description{
convert steepness from logit
}
