% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSR.R
\name{srrTMB}
\alias{srrTMB}
\title{Fits Stock Recruitment Relationships (SRR) in TBM}
\usage{
srrTMB(
  object,
  spr0,
  s = NULL,
  s.est = TRUE,
  s.logitsd = 1.3,
  plim = NULL,
  nyears = NULL,
  report.sR0 = FALSE,
  inits = NULL,
  lower = NULL,
  upper = NULL,
  SDreport = TRUE
)
}
\arguments{
\item{object}{Input FLSR object.}

\item{spr0}{unfished spawning biomass per recruit from FLCore::spr0(FLStock)}

\item{s}{steepness parameter of SRR (fixed or prior mean)}

\item{s.est}{option to estimate steepness}

\item{s.logitsd}{prior sd for logit(s), default is 1.3 (flat) if s.est = TRUE}

\item{plim}{determines the minimum break point of the hockey-stick as ratio blim/b0}

\item{nyears}{yearMeans from the tail used to compute a,b from the reference spr0 (default all years)}

\item{report.sR0}{option to report s and R0 instead of a,b}

\item{inits}{option to specify initial values of log(r0), log(SigR) and logit(s)}

\item{lower}{option to specify lower bounds of log(r0), log(SigR) and logit(s)}

\item{upper}{option to specify upper bounds of log(r0), log(SigR) and logit(s)}

\item{SDreport}{option to converge hessian and get vcov}
}
\value{
A list containing elements 'FLSR', of class \emph{FLSR}
}
\description{
Fits Stock Recruitment Relationships (SRR) in TBM
}
