% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSR.R
\name{srrTMB}
\alias{srrTMB}
\title{Fits Stock Recruitment Relationships (SRR) in TBM}
\usage{
srrTMB(
  object,
  spr0,
  s = NULL,
  s.est = TRUE,
  s.logitsd = 1.4,
  plim = 0.05,
  pmax = 0.5,
  nyears = NULL,
  report.sR0 = FALSE,
  inits = NULL,
  lower = NULL,
  upper = NULL,
  SDreport = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{object}{Input FLSR = as.FLSR(stock,model) object with current model options
\itemize{
\item bevholtSV
\item rickerSV
\item segreg
\item geomean
}}

\item{spr0}{unfished spawning biomass per recruit from FLCore::spr0(FLStock)}

\item{s}{steepness parameter of SRR (fixed or prior mean)}

\item{s.est}{option to estimate steepness}

\item{s.logitsd}{prior sd for logit(s), default is 1.4 (flat) if s.est = TRUE}

\item{plim}{determines the minimum break point of the hockey-stick as ratio blim/b0}

\item{pmax}{determines the maximum break point of the hockey-stick as ratio blim/b0}

\item{nyears}{yearMeans from the tail used to compute a,b from the reference spr0 (default all years)}

\item{report.sR0}{option to report s and R0 instead of a,b}

\item{inits}{option to specify initial values of log(r0), log(SigR) and logit(s)}

\item{lower}{option to specify lower bounds of log(r0), log(SigR) and logit(s)}

\item{upper}{option to specify upper bounds of log(r0), log(SigR) and logit(s)}

\item{SDreport}{option to converge hessian and get vcov}

\item{verbose}{if TRUE, it shows tracing}
}
\value{
A list containing elements 'FLSR', of class \emph{FLSR}
}
\description{
Fits Stock Recruitment Relationships (SRR) in TBM
}
\examples{
data(ple4)
gm <- srrTMB(as.FLSR(ple4,model=geomean),spr0=spr0y(ple4))
bh <- srrTMB(as.FLSR(ple4,model=bevholtSV),spr0=spr0y(ple4))
ri <- srrTMB(as.FLSR(ple4,model=rickerSV),spr0=spr0y(ple4))
hs <- srrTMB(as.FLSR(ple4,model=segreg),spr0=spr0y(ple4),plim=0.05,pmax=0.3)
srs = FLSRs(gm=gm,bh=bh,ri=ri,hs=hs) # combine
plotsrs(srs) 
plotsrts(srs) 
do.call(rbind,lapply(srs,AIC))
}
